removed references to catalogFunctions
